Squashfs: add corruption check in get_dir_index_using_offset()
We read the size (of the name) field from disk.  This value should
be sanity checked for correctness to avoid blindly reading
huge amounts of unnecessary data from disk on corruption.

Note, here we're not actually reading the name into a buffer, but
skipping it, and so corruption doesn't cause buffer overflow, merely
lots of unnecessary amounts of data to be read.

Change-Id: I7b21b9765be4687faf91c84ce7805dc1a6aeb925
Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>